% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poispkgs-package.R
\docType{package}
\name{poispkgs-package}
\alias{poispkgs}
\alias{poispkgs-package}
\title{poispkgs: Attach Poisson Packages}
\description{
Attaches CRAN and GitHub packages commonly used in Poisson analyses in the preferred order.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/poispkgs}
  \item Report bugs at \url{https://github.com/poissonconsulting/poispkgs}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
  \item Evan Amies-Galonski \email{evanamiesgalonski@gmail.com} (\href{https://orcid.org/0000-0003-1096-2089}{ORCID})
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
}

Other contributors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Nadine Hussein \email{nadine@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-4470-8361}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
