% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{poispkgs}
\alias{poispkgs}
\alias{package-poispkgs}
\alias{poispkgs-package}
\title{poispkgs}
\description{
Attaches CRAN and GitHub R packages used in Poison Consulting scripts in the preferred order.
}
\seealso{
\url{https://github.com/poissonconsulting/poispkgs}
}
