% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poispkgs_packages}
\alias{poispkgs_packages}
\title{List all packages in poispkgs}
\usage{
poispkgs_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include poispkgs in the list?}
}
\description{
List all packages in poispkgs
}
\examples{
poispkgs_packages()
}
