% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-github-pkgs.R
\name{update_github_pkgs}
\alias{update_github_pkgs}
\title{Update GitHub Packages}
\usage{
update_github_pkgs(force = FALSE)
}
\arguments{
\item{force}{A flag indicating whether to update even if the package has not changed since the last update.}
}
\value{
A character vector of the installed github packages.
}
\description{
Updates all packages that were installed from GitHub
}
\references{
Modified from \url{https://stackoverflow.com/questions/32538052/update-all-packages-from-github/35016988#35016988}
}
