% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poispkgs-package.R
\docType{package}
\name{poispkgs-package}
\alias{poispkgs}
\alias{poispkgs-package}
\title{poispkgs: Attach Poisson Packages}
\description{
Installs and attaches CRAN and Poisson drat packages commonly used in Poisson
    analyses in the preferred order.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/poispkgs}
  \item Report bugs at \url{https://github.com/poissonconsulting/poispkgs}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Evan Amies-Galonski \email{evan@poissonconsulting.ca} (\href{https://orcid.org/-3188}{ORCID})
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
}

Other contributors:
\itemize{
  \item Chris Muir [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
